# Improve performance by loading Swift tools version directly from Package manifests