--- a/Sources/TuistCore/Models/PackageSettings.swift
+++ b/Sources/TuistCore/Models/PackageSettings.swift
@@ -18,9 +18,6 @@ public struct PackageSettings: Equatable, Codable {
     /// The custom project options for each project generated from a swift package
     public let projectOptions: [String: XcodeGraph.Project.Options]
 
-    /// Swift tools version of the parsed `Package.swift`
-    public let swiftToolsVersion: Version
-
     /// Initializes a new `PackageSettings` instance.
     /// - Parameters:
     ///    - productTypes: The custom `Product` types to be used for SPM targets.
@@ -32,15 +29,13 @@ public struct PackageSettings: Equatable, Codable {
         productDestinations: [String: Destinations],
         baseSettings: Settings,
         targetSettings: [String: SettingsDictionary],
-        projectOptions: [String: XcodeGraph.Project.Options] = [:],
-        swiftToolsVersion: Version
+        projectOptions: [String: XcodeGraph.Project.Options] = [:]
     ) {
         self.productTypes = productTypes
         self.productDestinations = productDestinations
         self.baseSettings = baseSettings
         self.targetSettings = targetSettings
         self.projectOptions = projectOptions
-        self.swiftToolsVersion = swiftToolsVersion
     }
 }
 
@@ -51,16 +46,14 @@ public struct PackageSettings: Equatable, Codable {
             productDestinations: [String: Destinations] = [:],
             baseSettings: Settings = Settings.default,
             targetSettings: [String: SettingsDictionary] = [:],
-            projectOptions: [String: XcodeGraph.Project.Options] = [:],
-            swiftToolsVersion: Version = Version("5.4.9")
+            projectOptions: [String: XcodeGraph.Project.Options] = [:]
         ) -> PackageSettings {
             PackageSettings(
                 productTypes: productTypes,
                 productDestinations: productDestinations,
                 baseSettings: baseSettings,
                 targetSettings: targetSettings,
-                projectOptions: projectOptions,
-                swiftToolsVersion: swiftToolsVersion
+                projectOptions: projectOptions
             )
         }
     }