--- a/Tests/TuistLoaderTests/Loaders/PackageSettingsLoaderTests.swift
+++ b/Tests/TuistLoaderTests/Loaders/PackageSettingsLoaderTests.swift
@@ -85,8 +85,7 @@ final class PackageSettingsLoaderTests: TuistUnitTestCase {
                 ],
                 defaultSettings: .recommended
             ),
-            targetSettings: [:],
-            swiftToolsVersion: Version(stringLiteral: "5.4.9")
+            targetSettings: [:]
         )
         verify(manifestLoader)
             .register(plugins: .any)