… Tests/TuistLoaderTests/SwiftPackageManager/PackageInfoMapperTests.swift