Initiative: https://community.tuist.io/t/improve-tuists-performance/130

### Short description 📝

The `swift package tools-version` command is slow – and running that for each manifest, even on hot runs, adds up quickly.

Results using a real-world project with the current state of `main`:
```
hyperfine '"/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open' --warmup 1
Benchmark 1: "/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open
  Time (mean ± σ):      4.431 s ±  0.286 s    [User: 9.103 s, System: 3.453 s]
  Range (min … max):    4.188 s …  5.150 s    10 runs
```

Results with the changes from this PR:
```
hyperfine '"/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open' --warmup 1
Benchmark 1: "/Users/marekfort/Library/Developer/Xcode/DerivedData/TuistExt-fnvyzgrgowotpwgcllhszvqbqzxi/Build/Products/Release/tuist_ext" generate --no-open
  Time (mean ± σ):      3.965 s ±  0.057 s    [User: 8.326 s, System: 3.328 s]
  Range (min … max):    3.886 s …  4.063 s    10 runs
```

That is, this change leads to around ~12 % speedup on warm runs :tada:

### How to test the changes locally 🧐

- CI should pass. You can also play with any fixture that has SPM dependencies.

### Contributor checklist ✅

- [x] The code has been linted using run `mise run lint-fix`
- [x] The change is tested via unit testing or acceptance testing, or both
- [x] The title of the PR is formulated in a way that is usable as a changelog entry
- [x] In case the PR introduces changes that affect users, the documentation has been updated

### Reviewer checklist ✅

- [ ] The code architecture and patterns are consistent with the rest of the codebase
- [ ] Reviewer has checked that, if needed, the PR includes the label `changelog:added`, `changelog:fixed`, or `changelog:changed`, and the title is usable as a changelog entry